+++ tests/errors/nativeptr_err/nativeptr_err.c
@@ -10,15 +10,15 @@
 #error nativeptr_err.nativeptr_err:23:30 Expected two different pointer types but: one was Own<voidptr>, the other: nativeptr<'T>
 #error nativeptr_err.nativeptr_err:23:40 Expected a pointer type but was: Own<voidptr>
 #error nativeptr_err.nativeptr_err:27:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:28:15 Expected nativeptr, __outptr or voidptr type.
-#error nativeptr_err.nativeptr_err:31:15 Expected nativeptr, __outptr or voidptr type.
-#error nativeptr_err.nativeptr_err:33:3 Type mismatch: '__voidinptr' <> 'uint8'.
-#error nativeptr_err.nativeptr_err:34:3 Type mismatch: '__voidinptr' <> 'unit'.
+#error nativeptr_err.nativeptr_err:28:15 Expected nativeptr, OutPtr or voidptr type.
+#error nativeptr_err.nativeptr_err:31:15 Expected nativeptr, OutPtr or voidptr type.
+#error nativeptr_err.nativeptr_err:33:3 Type mismatch: 'VoidInPtr' <> 'uint8'.
+#error nativeptr_err.nativeptr_err:34:3 Type mismatch: 'VoidInPtr' <> 'unit'.
 #error nativeptr_err.nativeptr_err:38:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:39:15 Expected __inptr, __voidinptr or __outptr type.
-#error nativeptr_err.nativeptr_err:43:9 Expected __inptr, __voidinptr or __outptr type.
+#error nativeptr_err.nativeptr_err:39:15 Expected InPtr, VoidInPtr or OutPtr type.
+#error nativeptr_err.nativeptr_err:43:9 Expected InPtr, VoidInPtr or OutPtr type.
 #error nativeptr_err.nativeptr_err:52:14 Type mismatch: 'nativeptr<int>' <> 'int'.
-#error nativeptr_err.nativeptr_err:53:14 Type mismatch: '__inptr<int>' <> 'nativeptr<int>'.
+#error nativeptr_err.nativeptr_err:53:14 Type mismatch: 'InPtr<int>' <> 'nativeptr<int>'.
 #error nativeptr_err.nativeptr_err:54:14 Type mismatch: 'voidptr' <> 'nativeptr<int>'.
 #error nativeptr_err.nativeptr_err:63:11 Expected a variable.
 #error nativeptr_err.nativeptr_err:66:11 Expected a variable.
@@ -26,7 +26,7 @@
 #error nativeptr_err.nativeptr_err:72:24 Expected pointer type.
 #error nativeptr_err.nativeptr_err:78:15 This function misses some argument.
 #error nativeptr_err.nativeptr_err:79:20 Expected pointer type.
-#error nativeptr_err.nativeptr_err:80:25 Expected nativeptr but was __outptr.
+#error nativeptr_err.nativeptr_err:80:25 Expected nativeptr but was OutPtr.
 #error nativeptr_err.nativeptr_err:85:15 This function misses some argument.
 #error nativeptr_err.nativeptr_err:86:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:87:26 Expected nativeptr but was __inptr.
+#error nativeptr_err.nativeptr_err:87:26 Expected nativeptr but was InPtr.