+++ tests/primitives/native_cast/native_cast.fs
@@ -7,8 +7,8 @@ open Std.Ptr
 
 let main _ =
   let mutNull: voidptr = Ptr.nullPtr
-  let constNull: __voidinptr = __nativeCast mutNull
-  let constIntPtr: __inptr<int> = __nativeCast constNull
+  let constNull: VoidInPtr = __nativeCast mutNull
+  let constIntPtr: InPtr<int> = __nativeCast constNull
   let intPtr: nativeptr<int> = __nativeCast constIntPtr
 
   let address: nativeint = __nativeCast intPtr